+++ praw/models/reddit/collections.py
@@ -169,8 +169,8 @@ class CollectionModeration(PRAWBase):
         """Update the collection's display layout.
 
         :param display_layout: Either ``"TIMELINE"`` for events or discussions or
-            ``"GALLERY"`` for images or memes. Passing ``""`` or ``None`` will clear the
-            set layout and ``collection.display_layout`` will be ``None``, however, the
+            ``"GALLERY"`` for images or memes. Passing ``None`` will clear the set
+            layout and ``collection.display_layout`` will be ``None``, however, the
             collection will appear on Reddit as if ``display_layout`` is set to
             ``"TIMELINE"``.
 
@@ -441,7 +441,9 @@ class SubredditCollectionsModeration(PRAWBase):
         super().__init__(reddit, _data)
         self.subreddit_fullname = sub_fullname
 
-    def create(self, title: str, description: str):
+    def create(
+        self, title: str, description: str, display_layout: Optional[str] = None
+    ):
         """Create a new :class:`.Collection`.
 
         The authenticated account must have appropriate moderator permissions in the
@@ -449,6 +451,10 @@ class SubredditCollectionsModeration(PRAWBase):
 
         :param title: The title of the collection, up to 300 characters.
         :param description: The description, up to 500 characters.
+        :param display_layout: Either ``"TIMELINE"`` for events or discussions or
+            ``"GALLERY"`` for images or memes. Passing ``""`` or ``None`` will make the
+            collection appear on Reddit as if this is set to ``"TIMELINE"`` (default:
+            ``None``).
 
         :returns: The newly created :class:`.Collection`.
 
@@ -460,18 +466,31 @@ class SubredditCollectionsModeration(PRAWBase):
             new_collection = my_sub.collections.mod.create("Title", "desc")
             new_collection.mod.add_post("bgibu9")
 
+        To specify the display layout as ``"GALLERY"`` when creating the collection:
+
+        .. code-block:: python
+
+            my_sub = reddit.subreddit("test")
+            new_collection = my_sub.collections.mod.create(
+                title="Title", description="desc", display_layout="GALLERY"
+            )
+            new_collection.mod.add_post("bgibu9")
+
         .. seealso::
 
             :meth:`~.CollectionModeration.delete`
 
         """
+        data = {
+            "sr_fullname": self.subreddit_fullname,
+            "title": title,
+            "description": description,
+        }
+        if display_layout:
+            data["display_layout"] = display_layout
         return self._reddit.post(
             API_PATH["collection_create"],
-            data={
-                "sr_fullname": self.subreddit_fullname,
-                "title": title,
-                "description": description,
-            },
+            data=data,
         )