+++ CHANGES.rst
@@ -11,6 +11,8 @@ Unreleased
 - :meth:`.pin` to manage pinned submissions on the authenticated user's profile.
 - :meth:`.update_display_layout` to update the display layout of posts in a
   :class:`.Collection`.
+- :meth:`.SubredditCollectionsModeration.create` keyword argument ``display_layout`` for
+  specifying a display layout when creating a :class:`.Collection`.
 
 **Changed**